+++ src/components/context-menu/subcomponents/context-menu-separator.tsx
@@ -1,0 +1,11 @@
+/**
+ * コンテキストメニュー/サブメニューの区切り線を表す要素
+ */
+export function ContextMenuSeparator() {
+  return (
+    <div
+      className="border-t border-gray-200"
+      data-component-name="context-menu-separator"
+    />
+  );
+}
