+++ src/components/context-menu/subcomponents/submenu-context.tsx
@@ -1,0 +1,18 @@
+import { useState } from "react";
+import type { ReactNode } from "react";
+import {
+  submenuStateContext,
+  submenuStateSetterContext,
+} from "../contexts/submenu-context";
+
+export function SubmenuStateProvider({ children }: { children: ReactNode }) {
+  const [state, setState] = useState<"open" | "closed">("closed");
+
+  return (
+    <submenuStateContext.Provider value={{ state }}>
+      <submenuStateSetterContext.Provider value={{ setState }}>
+        {children}
+      </submenuStateSetterContext.Provider>
+    </submenuStateContext.Provider>
+  );
+}
